move grouped elements into subfolder
