+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/index.ts
@@ -1,0 +1,4 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export * from './Toolbar';
+export * from './types';
