+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/types.ts
@@ -1,7 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { MenuItemStyle } from '@frontify/fondue';
-import type { ReactNode } from 'react';
+import { type ReactNode } from 'react';
+
+import { type FlyoutToolbarItem, type ToolbarItem } from './Toolbar';
 
 export type BlockItemWrapperProps = {
     children: ReactNode;
@@ -14,39 +15,4 @@ export type BlockItemWrapperProps = {
     outlineOffset?: number;
     shouldBeShown?: boolean;
     showAttachments: boolean;
-};
-
-export type ToolbarProps = {
-    items: ToolbarItem[];
-    flyoutMenu: FlyoutState & { items: FlyoutToolbarItem[][] };
-    attachments: FlyoutState & { isEnabled: boolean };
-    isDragging?: boolean;
-};
-
-export type FlyoutState = {
-    isOpen: boolean;
-    onOpenChange: (isOpen: boolean) => void;
-};
-
-type BaseToolbarItem = {
-    icon: JSX.Element;
-    tooltip?: string;
-};
-
-type DraghandleToolbarItem = BaseToolbarItem & {
-    draggableProps: Record<string, unknown>;
-    setActivatorNodeRef?: (node: HTMLElement | null) => void;
-};
-
-type ButtonToolbarItem = BaseToolbarItem & {
-    onClick: () => void;
-};
-
-export type ToolbarItem = DraghandleToolbarItem | ButtonToolbarItem;
-
-export type FlyoutToolbarItem = {
-    title: string;
-    onClick: () => void;
-    icon: JSX.Element;
-    style?: MenuItemStyle;
 };
